move evacs to new module
