+++ bin/dipy_evac_plus
@@ -1,7 +1,7 @@
 #!python
 
 from dipy.workflows.flow_runner import run_flow
-from dipy.workflows.segment import EVACPlusFlow
+from dipy.workflows.nn import EVACPlusFlow
 
 if __name__ == "__main__":
     run_flow(EVACPlusFlow())
